+++ relm4/src/extensions/mod.rs
@@ -178,6 +178,7 @@ container_child_impl! {
 }
 
 #[cfg(feature = "libadwaita")]
+#[cfg_attr(docsrs, doc(cfg(feature = "libadwaita")))]
 mod libadwaita {
     use super::ContainerChild;