+++ relm4/src/extensions/remove.rs
@@ -36,6 +36,7 @@ impl RelmRemoveExt for gtk::FlowBox {
 }
 
 #[cfg(feature = "libadwaita")]
+#[cfg_attr(docsrs, doc(cfg(feature = "libadwaita")))]
 impl RelmRemoveExt for adw::PreferencesGroup {
     fn container_remove(&self, widget: &impl AsRef<Self::Child>) {
         use adw::prelude::PreferencesGroupExt;